+++ src/components/radio/editForm/Radio.edit.data.js
@@ -76,6 +76,26 @@ export default [
     },
   },
   {
+    type: 'select',
+    input: true,
+    label: 'Storage Type',
+    key: 'dataType',
+    clearOnHide: true,
+    tooltip: 'The type to store the data. If you select something other than autotype, it will force it to that type.',
+    weight: 12,
+    template: '<span>{{ item.label }}</span>',
+    dataSrc: 'values',
+    data: {
+      values: [
+        { label: 'Autotype', value: 'auto' },
+        { label: 'String', value: 'string' },
+        { label: 'Number', value: 'number' },
+        { label: 'Boolean', value: 'boolean' },
+        { label: 'Object', value: 'object' },
+      ],
+    },
+  },
+  {
     key: 'template',
     conditional: {
       json: { '===': [{ var: 'data.dataSrc' }, 'url'] },